Fetched default code from school into project
I also wanted to add typescript into the code but decided in the last minute against it because it will then be more stuff to learn in a short span of time. Maybe in the front-end I will try to import typescript together with vue class. That sounds quite neat.